+++ conts.go
@@ -1,0 +1,5 @@
+package regattaClock
+
+const (
+	zeroTime = "00:00.0"
+)
